big refactor
use ProcessVariable struct with ConnectionMonitor for each channel
which allows us to determine when PVs are connected and disconnected.
Also adjust how create_monitor works, and rename to set_monitor.
Monitors are actually created at object construction, and set_monitor
sets the variable (pointer) which is updated by the monitor