get values by type instead of dumping string
relying on the user to pass the correct type to the set_monitor call and
using that to pvget the value as the type instead of just dumping
everything to a string. The value field is still dumped to a string (as
opposed to getting as PVString) when the user passes a ref to
std::string to set monitor. Perhaps this can be useful if the type of the
underlying PV is not known or supported.